cleanup ofauthentication components
