+++ Client/.vscode/settings.json
@@ -25,5 +25,20 @@
     "editor.codeActionsOnSave": {
         "source.fixAll.eslint": true
     },
-    "todo-tree.regex.regexCaseSensitive": false
+    "todo-tree.regex.regexCaseSensitive": false,
+    "sqltools.connections": [
+        {
+            "mysqlOptions": {
+                "authProtocol": "default"
+            },
+            "previewLimit": 50,
+            "server": "databoomserver.mysql.database.azure.com",
+            "port": 4000,
+            "driver": "MySQL",
+            "name": "development",
+            "database": "databoom_db_test",
+            "username": "databoomuser@databoomserver",
+            "password": "Passw0rd"
+        }
+    ]
 }
