+++ Client/src/Components/Authentication/SignIn.tsx
@@ -1,21 +1,21 @@
-import React, { useState } from 'react';
-import Avatar from '@material-ui/core/Avatar';
-import Button from '@material-ui/core/Button';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox';
-import Link from '@material-ui/core/Link';
-import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
-import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
 import { FastField, Form, Formik } from 'formik'
-import { ISignInUser } from '../Models/Profile/IProfileModel';
-import { MuiTextFieldFormik } from '../Components/Forms/FormikFields';
-import { loginValidationSchema } from './AuthenticationValidationSchema';
 
+import Avatar from '@material-ui/core/Avatar'
+import Box from '@material-ui/core/Box'
+import Button from '@material-ui/core/Button'
+import Checkbox from '@material-ui/core/Checkbox'
+import Container from '@material-ui/core/Container'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import FormControlLabel from '@material-ui/core/FormControlLabel'
+import Grid from '@material-ui/core/Grid'
+import { ISignInUser } from '../../Models/Profile/IProfileModel'
+import Link from '@material-ui/core/Link'
+import LockOutlinedIcon from '@material-ui/icons/LockOutlined'
+import { MuiTextFieldFormik } from '../Forms/FormikFields'
+import React from 'react'
+import Typography from '@material-ui/core/Typography'
+import { loginValidationSchema } from './AuthenticationValidationSchema'
+import { makeStyles } from '@material-ui/core/styles'
 
 function Copyright() {
   return (
@@ -27,7 +27,7 @@ function Copyright() {
       {new Date().getFullYear()}
       {'.'}
     </Typography>
-  );
+  )
 }
 
 const useStyles = makeStyles((theme) => ({
@@ -48,10 +48,10 @@ const useStyles = makeStyles((theme) => ({
   submit: {
     margin: theme.spacing(3, 0, 2),
   },
-}));
+}))
 
 export default function SignIn() {
-  const classes = useStyles();
+  const classes = useStyles()
 
   const user: ISignInUser = {
     email: "",
@@ -87,7 +87,7 @@ export default function SignIn() {
               handleChange,
               handleBlur,
               handleSubmit
-            } = props;
+            } = props
             return (
               <>
                 <Form className={classes.form} noValidate>
@@ -149,5 +149,5 @@ export default function SignIn() {
         <Copyright />
       </Box>
     </Container>
-  );
+  )
 }
