+++ Client/src/Components/Authentication/SignUp.tsx
@@ -1,20 +1,20 @@
-import React, { Fragment, useState } from 'react';
-import Avatar from '@material-ui/core/Avatar';
-import Button from '@material-ui/core/Button';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import TextField from '@material-ui/core/TextField';
-import Link from '@material-ui/core/Link';
-import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
-import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import { FastField, Form, Formik } from 'formik';
-import { signupValidationSchema } from './AuthenticationValidationSchema';
-import { ISignUpUser } from '../Models/Profile/IProfileModel';
-import { MuiTextFieldFormik } from '../Components/Forms/FormikFields';
-import { DatePicker } from 'material-ui';
+import { FastField, Form, Formik } from 'formik'
+import React, { Fragment, useState } from 'react'
+
+import Avatar from '@material-ui/core/Avatar'
+import Box from '@material-ui/core/Box'
+import Button from '@material-ui/core/Button'
+import Container from '@material-ui/core/Container'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import Grid from '@material-ui/core/Grid'
+import { ISignUpUser } from '../../Models/Profile/IProfileModel'
+import Link from '@material-ui/core/Link'
+import LockOutlinedIcon from '@material-ui/icons/LockOutlined'
+import { MuiTextFieldFormik } from '../Forms/FormikFields'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import { makeStyles } from '@material-ui/core/styles'
+import { signupValidationSchema } from './AuthenticationValidationSchema'
 
 function Copyright() {
   return (
@@ -26,7 +26,7 @@ function Copyright() {
       {new Date().getFullYear()}
       {'.'}
     </Typography>
-  );
+  )
 }
 
 const useStyles = makeStyles((theme) => ({
@@ -47,12 +47,12 @@ const useStyles = makeStyles((theme) => ({
   submit: {
     margin: theme.spacing(3, 0, 2),
   },
-}));
+}))
 
 export default function SignUp() {
-  const classes = useStyles();
+  const classes = useStyles()
 
-  const [email, setEmail] = useState("");
+  const [email, setEmail] = useState("")
 
   const user: ISignUpUser = {
     firstName: "",
@@ -78,7 +78,7 @@ export default function SignUp() {
           </Avatar>
           <Typography component="h1" variant="h5">
             Sign up
-        </Typography>
+          </Typography>
           <Formik
             initialValues={user}
             validationSchema={signupValidationSchema}
@@ -93,7 +93,7 @@ export default function SignUp() {
                 handleChange,
                 handleBlur,
                 handleSubmit
-              } = props;
+              } = props
               return (
                 <>
                   <Form className={classes.form} noValidate>
@@ -198,7 +198,7 @@ export default function SignUp() {
                     </Grid>
                   </Form>
                 </>
-              );
+              )
             }}
           </Formik>
         </div>
@@ -207,6 +207,6 @@ export default function SignUp() {
         </Box>
       </Container>
     </Fragment >
-  );
+  )
 }
